requirements.txt
numpy==1.21.5
torch==1.11.0
torchvision==0.12.0
matplotlib==3.5.1
opencv-python==4.5.5.64
pillow==8.4.0
numba==0.54.1
loguru==0.5.3
matplotlib-inline==0.1.3
